Landed W3C page visibility module
The current issue is that "document.addEventListener(visibilitychange, function(){...})"
does not work if it is invoked in the document.DOMContentLoaded listeners, for example, at
JQuery(document).ready function.

The reason is that window.DOMContentLoaded is fired after the document.DOMContentLoaded.
Currently we have not found a proper event to override the document.addEventListener.
The beforeload is fine for window.addEventListener, while does not work for document.addEventListener
because the contentDocument will be reset after loading.